requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
PyPDF2==3.0.1
pytesseract==0.3.10
pdf2image==1.16.3
nltk==3.8.1
Pillow==10.0.0
